Add files via upload
Consider a scenario of your choice and implement CRUD (Create, Read,  Update and  Delete) operation in it along with the interfaces (use bootstrap).
Add New Students link as shown in task image should open new form with text fields First Name, Last Name, and Email Id. When the user submits the button, the form data should be saved into the students' table and the page should be redirected on the page as shown in image above. In the image above, the last column is showing three different icons (having links) for viewing, updating, and deleting the record of students (each row), when the user clicks on a particular icon, it should perform the specific CRUD operation on the table data. The image below is displaying the form for adding student records.